deleted DTO classes(found better way to do it), created a player class

+++ Backend/RM-Game/src/main/java/com/Philip/RMGame/data/enums/AnnualBonus.java
@@ -7,13 +7,13 @@ public enum AnnualBonus {
     TOOLS(50000),
     METAL(25000);
 
-    private final int value;
+    private final long value;
 
-    AnnualBonus(int value) {
+    AnnualBonus(long value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 }
